+++ lib/src/enums/card_rank.dart
@@ -18,6 +18,48 @@ enum CardRank {
 
   const CardRank(this.symbol, this.value);
 
+  factory CardRank.fromString(String rank) {
+    switch (rank.toUpperCase()) {
+      case '2':
+        return CardRank.two;
+      case '3':
+        return CardRank.three;
+      case '4':
+        return CardRank.four;
+      case '5':
+        return CardRank.five;
+      case '6':
+        return CardRank.six;
+      case '7':
+        return CardRank.seven;
+      case '8':
+        return CardRank.eight;
+      case '9':
+        return CardRank.nine;
+      case 'T':
+      case '10':
+        return CardRank.ten;
+      case 'J':
+        return CardRank.jack;
+      case 'Q':
+        return CardRank.queen;
+      case 'K':
+        return CardRank.king;
+      case 'A':
+        return CardRank.ace;
+      default:
+        throw ArgumentError('Unknown card rank: $rank');
+    }
+  }
+  int compareTo(CardRank other) {
+    return value.compareTo(other.value);
+  }
+
+  bool operator <(CardRank other) => value < other.value;
+  bool operator >(CardRank other) => value > other.value;
+  bool operator <=(CardRank other) => value <= other.value;
+  bool operator >=(CardRank other) => value >= other.value;
+
   @override
   String toString() => symbol;
 }