+++ CHANGELOG.md
@@ -121,3 +121,8 @@
 
 ## 1.1.26(2025-05-19)
 - added TournamentGameRecord.isNotEmpty
+
+## 1.1.26(2025-05-26)
+- added fromString method for enums: CardRank, CardSuit, ScoringType, Suit
+- migration from using deprecated Parser to GrammarParser
+- Card.fromString now uses CardRank.fromString and CardSuit.fromString